requirements.txt
Appium-Python-Client==5.0.0
selenium==4.26.0
urllib3==1.26.18
typing_extensions== 4.11.0
certifi==2025.6.15
mitmproxy==11.0.2
urwid==2.6.16
behave==1.3.3
pytest==  8.4.2
requests==2.31.0
